+++ getGradePoints.h
@@ -1,4 +1,4 @@
 #ifndef JAN_3_24_GETGRADEPOINTS_H
 #define JAN_3_24_GETGRADEPOINTS_H
-double getGradePoints(double numClass, double units);
+double getGradePoints(int numClass, double units);
 #endif
